backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
pydantic
requests
langchain-core
litellm
mcp
httpx

# Testing (run: pytest)
pytest
pytest-asyncio